+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.3.0 - 2023-04-XX
 
-- New feature: `processMarkup`
+- New feature: `processMarkup` and `processNestedMarkup`
 - New feature: `extractMarkup`
 
 ## 1.2.2 - 2023-03-16